+++ weather-booking-app-client/src/endpoint-caller/interfaces/bookings/BookingResponse.ts
@@ -1,0 +1,7 @@
+import { BookingDetails } from "./BookingDetails";
+import { WeatherOption } from "./WeatherOption";
+
+export interface BookingResponse {
+  booking: BookingDetails[];
+  weather_option: WeatherOption[];
+}
