+++ weather-booking-app-client/src/endpoint-caller/interfaces/bookings/WeatherOption.ts
@@ -1,0 +1,9 @@
+import { WeatherOptionChoice } from "../enums/WeatherOptionChoice";
+import { WeatherOptionType } from "../enums/WeatherOptionType";
+import { WeatherValueType } from "../enums/WeatherValueType";
+
+export interface WeatherOption {
+  option_type: WeatherOptionType;
+  option_name: WeatherOptionChoice;
+  value_type: WeatherValueType;
+}
