+++ weather-booking-app-client/src/endpoint-caller/interfaces/enums/WeatherOptionChoice.ts
@@ -1,0 +1,14 @@
+export type WeatherOptionChoice =
+  | "Sunny"
+  | "Rainy"
+  | "Cloudy"
+  | "Stormy"
+  | "Cool"
+  | "Mild"
+  | "Warm"
+  | "Hot"
+  | "Freezing"
+  | "No Wind"
+  | "Calm"
+  | "Windy"
+  | "Gusty";
