+++ weather-booking-app-client/src/endpoint-caller/interfaces/enums/WeatherOptionType.ts
@@ -1,0 +1,1 @@
+export type WeatherOptionType = "Weather" | "Temperature" | "Wind";
