+++ weather-booking-app-client/src/endpoint-caller/interfaces/enums/WeatherValueType.ts
@@ -1,0 +1,1 @@
+export type WeatherValueType = "Km/h" | "Celsius";
