+++ weather-booking-app-client/src/endpoint-caller/interfaces/locations/Location.ts
@@ -1,0 +1,6 @@
+export interface Location {
+  suburb: string;
+  state: string;
+  postcode: string;
+  country: string;
+}
